Add functionality for crud order

+++ src/main/java/com/intern/techtest/security/jwt/JwtUtils.java
@@ -1,19 +1,15 @@
 package com.intern.techtest.security.jwt;
 
-import java.security.Key;
 import java.util.Date;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
 import com.intern.techtest.security.services.UserDetailsImpl;
 
 import io.jsonwebtoken.*;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
 
 @Component
 public class JwtUtils {
@@ -23,7 +19,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${techtest.app.jwtExpirationMs}")
-    private int jwtExpirationMs;
+    private long jwtExpirationMs;
 
     public String generateJwtToken(UserDetailsImpl userPrincipal) {
         return generateTokenFromUsername(userPrincipal.getUsername());
